Adds new modes of operation.
Available modes are Dual DashBoard (default). Single full screen DashBoard, and Single Rotating DashBoard with additional static DashBoard. Mode is specified as an argument. Clients mode must match server. In the future, clients should get the info from the server, but this is enough for now just for the proof of concept.